Add Veo3 and Imagen4 snippets to README
