+++ Sources/AIProxy/EachAI/EachAICreatePredictionRequestBody.swift
@@ -22,52 +22,3 @@ public struct EachAICreatePredictionRequestBody<T: Encodable>: Encodable {
         self.version = version
     }
 }
-
-struct EachAIVeoInput: Encodable {
-    public let generateAudio: Bool
-    public let imageURL: URL
-    public let prompt: String
-
-    private enum CodingKeys: String, CodingKey {
-        case generateAudio = "generate_audio"
-        case imageURL = "image_url"
-        case prompt
-    }
-
-    public init(
-        generateAudio: Bool,
-        imageURL: URL,
-        prompt: String
-    ) {
-        self.generateAudio = generateAudio
-        self.imageURL = imageURL
-        self.prompt = prompt
-    }
-}
-
-public struct EachAIImagenInput: Encodable {
-    public let prompt: String
-
-    public let aspectRatio: String?
-    public let outputFormat: String?
-    public let safetyFilterLevel: String?
-
-    private enum CodingKeys: String, CodingKey {
-        case prompt
-        case aspectRatio = "aspect_ratio"
-        case outputFormat = "output_format"
-        case safetyFilterLevel = "safety_filter_level"
-    }
-
-    public init(
-        prompt: String,
-        aspectRatio: String? = nil,
-        outputFormat: String? = nil,
-        safetyFilterLevel: String? = nil
-    ) {
-        self.prompt = prompt
-        self.aspectRatio = aspectRatio
-        self.outputFormat = outputFormat
-        self.safetyFilterLevel = safetyFilterLevel
-    }
-}
